requirements.txt
streamlit
opencv-python
Pillow
numpy
tensorflow
keras